fix: correct positioning of color picker (#42)
This fixes the positioning of the color picker so that it is still positioned correctly when the bpmn-js diagram is not in the top left. This also works when there are multiple bpmn-js diagrams on the same page, which is my personal use-case. Previously if you had multiple diagrams, the color picker popup would appear in the top left one no matter where the pad entry was clicked.

Co-authored-by: Nico Rehwaldt <nico.rehwaldt@camunda.com>